+++ src/Token/index.ts
@@ -1,0 +1,3 @@
+export {default} from './Token'
+export {default as IssueLabelToken} from './IssueLabelToken'
+export {default as AvatarToken} from './AvatarToken'